pyproject.toml
[tool.poetry]
name = "playrix-test-task"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-api-python-client = "^2.145.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
gspread = "^6.1.2"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.5.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.4"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
